[build] Consistently use host_machine rather than target_machine
Meson uses the same terminology as Autotools, where you build on the
build system, producing binaries for the host system, and the target
system is only relevant when compiling compilers. This isn't a compiler,
so it's confusing to use target_machine (although mostly harmless since
target_machine defaults to being the same as host_machine).

Signed-off-by: Simon McVittie <smcv@collabora.com>