[build] Allow passing arbitrary Meson options to package-native.sh
This can be used with a --cross-file to cross-compile an i386 version
of dxvk-native on x86_64. In particular, native Linux games that run on
Steam Runtime 3 'sniper' will want this if they're still compiled for
i386, because the sniper SDK is only available as an x86_64 container
with secondary i386 support.

Signed-off-by: Simon McVittie <smcv@collabora.com>